inference/requirements.txt
# Core ML libraries
torch==2.1.0
torchvision==0.16.0

# Diffusers and dependencies (compatible versions)
diffusers==0.21.4
transformers==4.35.2
accelerate==0.25.0
huggingface-hub==0.19.4

# ControlNet
controlnet-aux==0.0.7

# Image processing
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.24.3

# API
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Utilities
pydantic==2.5.3